Make MANAGE_DOCUMENTS permission optional
for those who can not or do not want to use it

Fixes #58
