+++ README.md
@@ -26,11 +26,11 @@ It uses the same internal APIs as `adb backup` which is deprecated and thus need
 ## Permissions
 * `android.permission.BACKUP` to back up application data.
 * `android.permission.ACCESS_NETWORK_STATE` to check if there is internet access when network storage is used.
-* `android.permission.MANAGE_DOCUMENTS` to retrieve the available storage roots.
 * `android.permission.MANAGE_USB` to access the serial number of USB mass storage devices.
 * `android.permission.WRITE_SECURE_SETTINGS` to change system backup settings and enable call log backup.
 * `android.permission.QUERY_ALL_PACKAGES` to get information about all installed apps for backup.
 * `android.permission.INSTALL_PACKAGES` to re-install apps when restoring from backup.
+* `android.permission.MANAGE_DOCUMENTS` to retrieve the available storage roots (optional) for better UX.
 
 ## Contributing
 Bug reports and pull requests are welcome on GitHub at https://github.com/stevesoltys/seedvault.
