+++ app/src/main/java/com/stevesoltys/seedvault/ui/storage/StorageActivity.kt
@@ -1,7 +1,16 @@
 package com.stevesoltys.seedvault.ui.storage
 
+import android.Manifest.permission.MANAGE_DOCUMENTS
+import android.content.Context
+import android.content.Intent
+import android.content.Intent.FLAG_GRANT_PERSISTABLE_URI_PERMISSION
+import android.content.Intent.FLAG_GRANT_READ_URI_PERMISSION
+import android.content.Intent.FLAG_GRANT_WRITE_URI_PERMISSION
+import android.content.pm.PackageManager.PERMISSION_GRANTED
+import android.net.Uri
 import android.os.Bundle
 import android.util.Log
+import androidx.activity.result.contract.ActivityResultContracts.OpenDocumentTree
 import androidx.annotation.CallSuper
 import androidx.appcompat.app.AlertDialog
 import com.stevesoltys.seedvault.R
@@ -16,6 +25,24 @@ private val TAG = StorageActivity::class.java.name
 class StorageActivity : BackupActivity() {
 
     private lateinit var viewModel: StorageViewModel
+
+    /**
+     * The official way to get a SAF [Uri] which we only use if we don't have the
+     * [MANAGE_DOCUMENTS] permission (via [canUseStorageRootsFragment]).
+     */
+    private val openDocumentTree = registerForActivityResult(OpenPersistableDocumentTree()) { uri ->
+        if (uri != null) {
+            Log.e(TAG, "OpenDocumentTree: $uri")
+            val authority = uri.authority ?: throw AssertionError("No authority in $uri")
+            val storageRoot = StorageRootResolver.getStorageRoots(this, authority).getOrNull(0)
+            if (storageRoot == null) {
+                viewModel.onUriPermissionResultReceived(null)
+            } else {
+                viewModel.onStorageRootChosen(storageRoot)
+                viewModel.onUriPermissionResultReceived(uri)
+            }
+        }
+    }
 
     @CallSuper
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -47,7 +74,11 @@ class StorageActivity : BackupActivity() {
         })
 
         if (savedInstanceState == null) {
-            showFragment(StorageRootsFragment.newInstance(isRestore()))
+            if (canUseStorageRootsFragment()) {
+                showFragment(StorageRootsFragment.newInstance(isRestore()))
+            } else {
+                openDocumentTree.launch(null)
+            }
         }
     }
 
@@ -61,13 +92,28 @@ class StorageActivity : BackupActivity() {
 
     private fun onInvalidLocation(errorMsg: String) {
         if (viewModel.isRestoreOperation) {
-            supportFragmentManager.popBackStack()
-            AlertDialog.Builder(this)
+            val dialog = AlertDialog.Builder(this)
                 .setTitle(getString(R.string.restore_invalid_location_title))
                 .setMessage(errorMsg)
                 .setPositiveButton(android.R.string.ok) { dialog, _ -> dialog.dismiss() }
-                .show()
+            if (canUseStorageRootsFragment()) {
+                // We have permission to use StorageRootsFragment,
+                // so pop the back stack to show it again
+                supportFragmentManager.popBackStack()
+                dialog.setPositiveButton(android.R.string.ok) { d, _ -> d.dismiss() }
+            } else {
+                // We don't have permission to use StorageRootsFragment,
+                // so give option to choose again or cancel.
+                dialog.setPositiveButton(android.R.string.ok) { _, _ ->
+                    openDocumentTree.launch(null)
+                }
+                dialog.setNegativeButton(android.R.string.cancel) { _, _ ->
+                    finishAfterTransition()
+                }
+            }
+            dialog.show()
         } else {
+            // just show error message, if this isn't restore
             showFragment(StorageCheckFragment.newInstance(getCheckFragmentTitle(), errorMsg))
         }
     }
@@ -80,10 +126,24 @@ class StorageActivity : BackupActivity() {
         return intent?.getBooleanExtra(INTENT_EXTRA_IS_SETUP_WIZARD, false) ?: false
     }
 
+    private fun canUseStorageRootsFragment(): Boolean {
+        return checkSelfPermission(MANAGE_DOCUMENTS) == PERMISSION_GRANTED
+    }
+
     private fun getCheckFragmentTitle() = if (viewModel.isRestoreOperation) {
         getString(R.string.storage_check_fragment_restore_title)
     } else {
         getString(R.string.storage_check_fragment_backup_title)
     }
 
+}
+
+private class OpenPersistableDocumentTree : OpenDocumentTree() {
+    override fun createIntent(context: Context, input: Uri?): Intent {
+        return super.createIntent(context, input).apply {
+            val flags = FLAG_GRANT_PERSISTABLE_URI_PERMISSION or
+                FLAG_GRANT_READ_URI_PERMISSION or FLAG_GRANT_WRITE_URI_PERMISSION
+            addFlags(flags)
+        }
+    }
 }
