+++ whitelist_com.stevesoltys.seedvault.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <config>
     <backup-transport-whitelisted-service
-        service="com.stevesoltys.seedvault/.transport.ConfigurableBackupTransportService"/>
+        service="com.stevesoltys.seedvault/.transport.ConfigurableBackupTransportService" />
+    <hidden-api-whitelisted-app package="com.stevesoltys.seedvault" />
 </config>
